+++ src/main/java/com/team8/kanban/domain/card/repository/CardRepository.java
@@ -1,6 +1,7 @@
-package com.team8.kanban.domain.card;
+package com.team8.kanban.domain.card.repository;
 
 
+import com.team8.kanban.domain.card.entity.Card;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CardRepository extends JpaRepository<Card, Long>, CardRepositoryCustom {
