+++ src/test/java/com/team8/kanban/card/CardServiceImplTest.java
@@ -1,8 +1,9 @@
 //package com.team8.kanban.card;
 //
-//import com.team8.kanban.domain.card.Card;
-//import com.team8.kanban.domain.card.CardRepository;
-//import com.team8.kanban.domain.card.CardServiceImpl;
+//
+//import com.team8.kanban.domain.card.entity.Card;
+//import com.team8.kanban.domain.card.repository.CardRepository;
+//import com.team8.kanban.domain.card.service.CardServiceImpl;
 //import com.team8.kanban.domain.card.dto.CardResponse;
 //import com.team8.kanban.domain.card.dto.CreateCardRequest;
 //import com.team8.kanban.domain.card.dto.UpdateCardRequest;
